printf

+++ printf_utils/ft_putstr.c
@@ -6,15 +6,15 @@
 /*   By: cdoria <cdoria@student.21-school.ru>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/15 23:03:47 by cdoria            #+#    #+#             */
-/*   Updated: 2021/11/15 23:28:12 by cdoria           ###   ########.fr       */
+/*   Updated: 2021/11/15 23:56:49 by cdoria           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../ft_printf.h"
 
-size_t	ft_strlen(const char *s)
+int	ft_strlen(const char *s)
 {
-	size_t	i;
+	int	i;
 
 	i = 0;
 	while (s[i] != 0)
